[New multicolumn] Spin in RenderMultiColumnSet::repaintFlowThreadContent() https://bugs.webkit.org/show_bug.cgi?id=132884
Reviewed by Beth Dakin.

On iOS, wikipedia pages can hang under RenderMultiColumnSet::repaintFlowThreadContent().
It appears that computedColumnHeight is set to 0 in prepareForLayout, but layout never happens
on the RenderMultiColumnSet in some cases, leaving the column height set to zero.
This caused columnIndexAtOffset() to return bad values, which resulted in very long loops
in repaintFlowThreadContent().

This fix is a stop-gap.

* rendering/RenderMultiColumnSet.cpp:
(WebCore::RenderMultiColumnSet::columnIndexAtOffset):

Identifier: 150993@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168882 268f45cc-cd09-0410-ab3c-d52691b4dbfc