WebKit2 View Gestures (Zoom): Document size changes during zoom gesture cause erratic scrolling https://bugs.webkit.org/show_bug.cgi?id=132944 <rdar://problem/15769454>
Reviewed by Simon Fraser.

* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::flushLayers):
If a layer flush occurs and we have an active transient zoom,
re-apply the transient zoom to the relevant platform layers, because
WebCore might have overwritten the relevant layer properties.

(WebKit::TiledCoreAnimationDrawingArea::applyTransientZoomToPage):
Reset m_transientZoomScale before flushLayers() so that we don't
re-apply the transient zoom when we're actually trying to commit it.


Identifier: 151011@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168904 268f45cc-cd09-0410-ab3c-d52691b4dbfc