[WK2] Add some missing state reset on crash https://bugs.webkit.org/show_bug.cgi?id=133036
Reviewed by Sam Weinig.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::resetState):
The missing reset for m_isTrackingTouchEvents would cause the WebPageProxy to send
touch events when it shouldn't. That would only be for an active touch sequence.

The missing reset m_lastVisibleContentRectUpdate would prevent pushing the new UI state
to the new page on reload, that is very bad.

Also moved m_videoFullscreenManager to be in the same order as initialization, and invalidate
the callbacks for dictation.


Identifier: 151108@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169066 268f45cc-cd09-0410-ab3c-d52691b4dbfc