REGRESSION: Standalone images need to take topContentInset into account https://bugs.webkit.org/show_bug.cgi?id=133092 -and corresponding- <rdar://problem/16945791>
Reviewed by Darin Adler.

The FrameView’s size is accurate in terms of the NSWindow size, but that is not
really what the ImageDocument code is looking for here. The ImageDocument wants to
know the size of the unobscured viewport area. We should use visibleSize() for
that.
* html/ImageDocument.cpp:
(WebCore::ImageDocument::scale):
(WebCore::ImageDocument::imageFitsInWindow):
(WebCore::ImageDocument::imageClicked):



Identifier: 151116@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169084 268f45cc-cd09-0410-ab3c-d52691b4dbfc