REGRESSION (WebKit2): Keyboard should have Search button in duckduckgo.com. https://bugs.webkit.org/show_bug.cgi?id=133183 <rdar://problem/17004207>
Reviewed by Geoff Garen.


Source/WebCore:
Adding new export.

* WebCore.exp.in:

Source/WebKit2:
The keyboard type should be search if the input type is search
or if it is text and it is inside a form with an action and either
name, id or title contain the word search. This matches the
heuristics we have in WK1 for iOS.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::getAssistedNodeInformation):



Identifier: 151199@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169207 268f45cc-cd09-0410-ab3c-d52691b4dbfc