Do not force a layout when changing the FixedLayoutSize in WebPage https://bugs.webkit.org/show_bug.cgi?id=133167
Patch by Benjamin Poulain <bpoulain@apple.com> on 2014-05-22
Reviewed by Andreas Kling.

WebPage::setFixedLayoutSize was forcing a layout immediately after changing the layout size on FrameView.
This is forcing plenty of extra layouts since the fixed layout size is one of many properties updated on the FrameView.

FrameView marks that layout is needed. It is not clear why a layout was performed immediately.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::setFixedLayoutSize):


Identifier: 151213@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169239 268f45cc-cd09-0410-ab3c-d52691b4dbfc