Remove invalid ASSERT in wordRangeForPosition. https://bugs.webkit.org/show_bug.cgi?id=133232
Reviewed by Ryosuke Niwa.

In WK2 for iOS this function is called on every selection change.
The selection could be set by Javascript in an element that is not visible.
In that case we won't be able to create a VisiblePosition therefore wordRangeForPosition
should not ASSERT that the VisiblePosition is not null, but rather
perform an early return if it is null.

* editing/VisibleUnits.cpp:
(WebCore::wordRangeFromPosition):



Identifier: 151246@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169280 268f45cc-cd09-0410-ab3c-d52691b4dbfc