Minor TileCoverageMap refinements https://bugs.webkit.org/show_bug.cgi?id=133468
Reviewed by Simon Fraser.

* platform/graphics/ca/mac/TileController.mm:
(WebCore::TileController::setContentsScale):
* platform/graphics/ca/mac/TileCoverageMap.h:
* platform/graphics/ca/mac/TileCoverageMap.mm:
(WebCore::TileCoverageMap::TileCoverageMap):
(WebCore::TileCoverageMap::setDeviceScaleFactor):
Keep the TileCoverageMap's contentsScale at the device scale factor, so that it
looks sharp on non-1x displays.

(WebCore::TileCoverageMap::update):
Decrease the height available for the TileCoverageMap by the topContentInset;
it's already being vertically offset by the topContentInset, but the height was
not adjusted, so it would run off the bottom of the window.

* platform/graphics/ca/mac/TileGrid.mm:
(WebCore::TileGrid::drawTileMapContents):
If using aggressive tile retention (and thus not coloring tiles based on age),
dim tiles that aren't in the visible cohort.


Identifier: 151503@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169563 268f45cc-cd09-0410-ab3c-d52691b4dbfc