REGRESSION: [CSS Regions] Regions with overflow:auto generate scrollbars even if the content flows into the following regions and as such, should not be scrolled https://bugs.webkit.org/show_bug.cgi?id=133481
Reviewed by Antti Koivisto.

Source/WebCore:
When enlarging the layout overflow of a scrollable region, the portion of the content that flows
in the region must be taken into consideration. If the content continues to flow into the next
region, the layout overflow of the current region must not be enlarged.

Test: fast/regions/scrollable-first-region.html

* rendering/RenderFlowThread.cpp:
(WebCore::RenderFlowThread::addRegionsOverflowFromChild):

LayoutTests:
Added test for a large item flowed through two regions with the first region having overflow:auto.
The region must NOT scroll.

* fast/regions/scrollable-first-region-expected.html: Added.
* fast/regions/scrollable-first-region.html: Added.


Identifier: 151518@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169586 268f45cc-cd09-0410-ab3c-d52691b4dbfc