[Mac] Hang when painting media before compositing mode activated https://bugs.webkit.org/show_bug.cgi?id=133655
Reviewed by Eric Carlson.

When asked to paint(), MediaPlayerPrivateAVFoundationObjC should not create an image generator
or a video output. These operations are synchronous, and will block the drawing thread and cause
a spin. paint() as opposed to paintImageInContext() is only called when painting the layer tree
and MediaPlayerPrivateAVFoundationObjC will paint by creating a AVPlayerLayer, not by painting
frames directly into the tree.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::paint):


Identifier: 151634@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169720 268f45cc-cd09-0410-ab3c-d52691b4dbfc